WebSockets/tests/submissions/Opera/cookies/005.html

<!doctype html>
<title>WebSockets: setting HttpOnly cookies in ws response, checking ws request</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setup({timeout:15100,
explicit_done:true})

var t = async_test(function(t) {
  var iframe = document.createElement('iframe');
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/set-cookie?'+encodeURIComponent('ws_test=test; Path=/; HttpOnly'));
  ws.onopen = t.step_func(function(e) {
    var ws2 = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo-cookie');
    ws2.onmessage = t.step_func(function(e) {
      ws.close();
      ws2.close();
      assert_regexp_match(/ws_test=test/, e.data);
      t.done();
    });
  });
  ws.onclose = t.step_func(function() {assert_unreached()});

  add_result_callback(function() {
    //This might be racy if a completion callback is being used to load multiple tests
    iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=; Path=/; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT');
    iframe.onload = done;
  });
  document.body.appendChild(iframe);
})
</script>
